Move aes and key-derivation code into a common package
