+++ internal/crypto/key_derivation.go
@@ -1,4 +1,4 @@
-package format
+package crypto
 
 import (
 	"crypto/sha256"
@@ -8,9 +8,9 @@ import (
 )
 
 // DeriveKeyFromMasterKey computes a key for a specific purpose and length using HKDF based on the master key.
-func DeriveKeyFromMasterKey(masterKey, uniqueID, purpose []byte, length int) []byte {
+func DeriveKeyFromMasterKey(masterKey, salt, purpose []byte, length int) []byte {
 	key := make([]byte, length)
-	k := hkdf.New(sha256.New, masterKey, uniqueID, purpose)
+	k := hkdf.New(sha256.New, masterKey, salt, purpose)
 
 	if _, err := io.ReadFull(k, key); err != nil {
 		panic("unable to derive key from master key, this should never happen")
